+++ src/main/java/com/maestria/gestionSolicitudes/domain/ActividadesRealizadasPracticaDocente.java
@@ -24,7 +24,7 @@ public class ActividadesRealizadasPracticaDocente {
     @Id
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "id_solicitud")
@@ -35,7 +35,7 @@ public class ActividadesRealizadasPracticaDocente {
     private SubTiposSolicitud subTiposSolicitud;
 
     @Column(name = "intensidad_horaria")
-    private Integer intensidadHoraria;
+    private Long intensidadHoraria;
 
     @Column(name = "horas_reconocer")
     private BigDecimal horasReconocer;
@@ -54,3 +54,8 @@ public class ActividadesRealizadasPracticaDocente {
     @Convert(converter = BooleanConverter.class)
     private Boolean aprobadoConcejo;
 }
+
+
+
+
+
