+++ src/main/java/com/maestria/gestionSolicitudes/domain/AdicionarAsignatura.java
@@ -19,9 +19,14 @@ public class AdicionarAsignatura {
     @Id
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "id_solicitud")
     private Solicitudes solicitud;
 }
+
+
+
+
+
