+++ src/main/java/com/maestria/gestionSolicitudes/domain/ApoyoEconomicoPublicacionEvento.java
@@ -22,7 +22,7 @@ public class ApoyoEconomicoPublicacionEvento {
     @Id
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "id_solicitud")
@@ -41,7 +41,7 @@ public class ApoyoEconomicoPublicacionEvento {
     private LocalDate fechaFin;
 
     @Column(name = "director_grupo_inv")
-    private Integer idDirector;
+    private Long idDirector;
 
     @Column(name = "titulo_publicacion")
     private String tituloPublicacion;
@@ -79,3 +79,8 @@ public class ApoyoEconomicoPublicacionEvento {
     @Column(name = "lufar_evento")
     private String lugarEvento;
 }
+
+
+
+
+
