+++ src/main/java/com/maestria/gestionSolicitudes/domain/DocumentosAvalPasantia.java
@@ -19,7 +19,7 @@ public class DocumentosAvalPasantia {
     @Id
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "id_aval_pasantia")
@@ -28,3 +28,8 @@ public class DocumentosAvalPasantia {
     @Column(name = "documento")
     private String documento;
 }
+
+
+
+
+
