+++ src/main/java/com/maestria/gestionSolicitudes/domain/HistorialEstadoSolicitudes.java
@@ -23,7 +23,7 @@ public class HistorialEstadoSolicitudes {
     @Id
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "id_solicitud")
@@ -49,3 +49,8 @@ public class HistorialEstadoSolicitudes {
     private Integer usuarioCreacion;
     
 }
+
+
+
+
+
