+++ src/main/java/com/maestria/gestionSolicitudes/dto/client/AsignaturaExternaDto.java
@@ -8,7 +8,12 @@ public class AsignaturaExternaDto {
     private String institutoProcedencia;
     private String nombreAsignatura;
     private Integer numeroCreditos;
-    private Integer intensidadHoraria;
+    private Long intensidadHoraria;
     private Double calificacion;
     private String contenidoProgramatico;
 }
+
+
+
+
+
