+++ src/main/java/com/maestria/gestionSolicitudes/dto/client/InformacionPersonalDto.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class InformacionPersonalDto {
-    private Integer id;
+    private Long id;
     private String nombres;
     private String apellidos;
     private String correo;
@@ -19,3 +19,8 @@ public class InformacionPersonalDto {
         return nombres + " " + apellidos;
     }
 }
+
+
+
+
+
