+++ src/main/java/com/maestria/gestionSolicitudes/dto/rest/request/ApoyoEconomicoPublicacionEventoRequest.java
@@ -11,7 +11,7 @@ public class ApoyoEconomicoPublicacionEventoRequest {
     private String tipoEvento;
     private String fechaInicio;
     private String fechaFin;
-    private Integer idDirectorGrupo;
+    private Long idDirectorGrupo;
     private String nombreDirectorGrupo;
     private String tituloPublicacion;
     private BigDecimal valorApoyo;
@@ -27,3 +27,8 @@ public class ApoyoEconomicoPublicacionEventoRequest {
     private String lugarEvento;
     private List<String> documentosAdjuntos;
 }
+
+
+
+
+
