+++ src/main/java/com/maestria/gestionSolicitudes/dto/rest/request/AprobarAsignaturaRequest.java
@@ -4,9 +4,14 @@ import lombok.Data;
 
 @Data
 public class AprobarAsignaturaRequest {
-    private Integer idAsignatura;
+    private Long idAsignatura;
     private String nombre;
     private String grupo;
     private String nombreDocente;
     private Boolean aprobado;
 }
+
+
+
+
+
