+++ src/main/java/com/maestria/gestionSolicitudes/dto/rest/request/DatosAvalarSolicitudDto.java
@@ -4,8 +4,13 @@ import lombok.Data;
 
 @Data
 public class DatosAvalarSolicitudDto {
-    private Integer idSolicitud;
+    private Long idSolicitud;
     private Boolean firmaTutor;
     private Boolean firmaDirector;
     private String documentoPdfSolicitud;
 }
+
+
+
+
+
