+++ src/main/java/com/maestria/gestionSolicitudes/repository/AdicionarAsignaturaRepository.java
@@ -6,6 +6,11 @@ import com.maestria.gestionSolicitudes.domain.AdicionarAsignatura;
 import com.maestria.gestionSolicitudes.domain.Solicitudes;
 
 
-public interface AdicionarAsignaturaRepository extends JpaRepository<AdicionarAsignatura, Integer> {
+public interface AdicionarAsignaturaRepository extends JpaRepository<AdicionarAsignatura, Long> {
     AdicionarAsignatura findBySolicitud(Solicitudes solicitud);
 }
+
+
+
+
+
