+++ src/main/java/com/maestria/gestionSolicitudes/repository/ApoyoEconomicoInvestigacionRepository.java
@@ -5,6 +5,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.maestria.gestionSolicitudes.domain.ApoyoEconomicoInvestigacion;
 import com.maestria.gestionSolicitudes.domain.Solicitudes;
 
-public interface ApoyoEconomicoInvestigacionRepository extends JpaRepository<ApoyoEconomicoInvestigacion, Integer> {
+public interface ApoyoEconomicoInvestigacionRepository extends JpaRepository<ApoyoEconomicoInvestigacion, Long> {
     ApoyoEconomicoInvestigacion findBySolicitud(Solicitudes solicitud);
 }
+
+
+
+
+
